notification

+++ Client/src/components/Admin/ProfileAdmin/Management/ManagementModal.tsx
@@ -47,15 +47,15 @@ const ManagementModal: FC<ManagersModalProps> = ({
   const dispatch = useAppDispatch();
 
   const managerToSave: IManager = editedManager
-  ? editedManager
-  : {
-      id: 0,
-      lastName: '',
-      firstName: '',
-      middleName: '',
-      email: '',
-      phone: '',
-    };
+    ? editedManager
+    : {
+        id: 0,
+        lastName: '',
+        firstName: '',
+        middleName: '',
+        email: '',
+        phone: '',
+      };
 
   useEffect(() => {
     console.log('Данные менеджера:', editedManager);
@@ -75,25 +75,36 @@ const ManagementModal: FC<ManagersModalProps> = ({
 
   const handleFormSubmit = (e: React.FormEvent) => {
     e.preventDefault();
-    try {
-      if (isAddingMode) {
-        onSaveAdd(managerToSave);
-      } else {
-        onSaveEdit(managerToSave);
+    const isConfirmed = window.confirm(
+      'Вы уверены, что хотите внести изменения?'
+    );
+    if (isConfirmed) {
+      try {
+        if (isAddingMode) {
+          onSaveAdd(managerToSave);
+        } else {
+          onSaveEdit(managerToSave);
+        }
+      } catch (error) {
+        console.error('Произошла ошибка при сохранении:', error);
       }
-    } catch (error) {
-      console.error('Произошла ошибка при сохранении:', error);
     }
   };
 
-  const handleDelete = () => {
-    //console.log('editedManager', editedManager);
-
-    if (editedManager && editedManager.id) {
+  const handleDelete = async () => {
+    const isConfirmed = window.confirm(
+      'Вы уверены, что хотите удалить данного менеджера?'
+    );
+    if (isConfirmed && editedManager && editedManager.id) {
       const managerId = editedManager.id;
 
-      dispatch(deleteManager({ managerId }));
-      onCloseEditModal();
+      try {
+        await dispatch(deleteManager({ managerId }));
+        // Если операция удаления успешна, закройте модальное окно
+        onCloseEditModal();
+      } catch (error) {
+        console.error('Произошла ошибка при удалении:', error);
+      }
     }
   };
